+++ src/components/domain/survey/index.ts
@@ -1,2 +1,4 @@
-export { default as TypeOfMeeting } from './TypeOfMeetingBox';
+export { default as TypeOfMeetingBox } from './TypeOfMeetingBox';
+export { default as GenderBox } from './GenderBox';
+export { default as AgeBox } from './AgeBox';
 export { default as SurveyTemplate } from './SurveyTemplate';
